renamed module sync_shared to rfidcards
