+++ resources/default-settings/jukebox.default.yaml
@@ -18,7 +18,7 @@ modules:
     host: hostif.linux
     bluetooth_audio_buttons: controls.bluetooth_audio_buttons
     gpio: gpio.gpioz.plugin
-    sync_shared: synchronisation.sync_shared
+    sync_rfidcards: synchronisation.rfidcards
   others:
   - music_cover_art
   - misc
@@ -143,6 +143,6 @@ speaking_text:
   speak_punct: False
   # Must be one of: female, male, croak, whisper
   voice: female
-sync_shared:
+sync_rfidcards:
   enable: false
-  config_file: ../../shared/settings/sync_shared.yaml
+  config_file: ../../shared/settings/sync_rfidcards.yaml
