+++ resources/default-settings/sync_rfidcards.default.yaml
