+++ src/webapp/public/locales/en/translation.json
@@ -22,8 +22,8 @@
           "timer_stop_player": "Stop player",
           "timer_fade_volume": "Fade volume",
           "toggle_output": "Toggle audio output",
-          "sync_shared_all": "Sync all audiofiles",
-          "sync_shared_change_on_rfid_scan": "Change activation of 'on RFID scan'"
+          "sync_rfidcards_all": "Sync all audiofiles and card entries",
+          "sync_rfidcards_change_on_rfid_scan": "Change activation of 'on RFID scan'"
         }
       },
       "controls-selector": {
@@ -63,10 +63,12 @@
           "description": "Choose the amount of minutes you want the action to be performed."
         },
         "synchronisation": {
+          "rfidcards": {
             "description": "Choose the state to set?",
             "label-toggle": "Toggle",
             "label-enable": "Enable",
             "label-disable": "Disable"
+          }
         }
       }
     },
