+++ src/webapp/src/commands/index.js
@@ -239,13 +239,13 @@ const commands = {
   },
 
   // Synchronisation
-  'sync_shared_all': {
-    _package: 'sync_shared',
+  'sync_rfidcards_all': {
+    _package: 'sync_rfidcards',
     plugin: 'ctrl',
     method: 'sync_all'
   },
-  'sync_shared_change_on_rfid_scan': {
-    _package: 'sync_shared',
+  'sync_rfidcards_change_on_rfid_scan': {
+    _package: 'sync_rfidcards',
     plugin: 'ctrl',
     method: 'sync_change_on_rfid_scan',
     argKeys: ['option']
