+++ src/webapp/src/components/Cards/controls/actions/synchronisation/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import CommandSelector from '../../command-selector';
-import ChangeOnRfidScan from './change-on-rfid-scan-options';
+import ChangeOnRfidScan from './rfidcards/change-on-rfid-scan-options';
 
 import { getActionAndCommand } from '../../../utils';
 
@@ -17,7 +17,7 @@ const SelectSynchronisation = ({
         actionData={actionData}
         handleActionDataChange={handleActionDataChange}
       />
-      {command === 'sync_shared_change_on_rfid_scan' &&
+      {command === 'sync_rfidcards_change_on_rfid_scan' &&
         <ChangeOnRfidScan
           actionData={actionData}
           handleActionDataChange={handleActionDataChange}
